GameRecommendations now use latest model.
There was an issue with using the latest model.  Even though we had a
system for updating the model as we went along, it was not being
propagated to the GameRecommendations, so they still looked like shit.
This update ensures that GameRecommendations are using the latest model
and already shows a remarkable improvement in the recommendations.